Chracter Encodings
The book said that I could use %lu to print large characters. I didn't really
understand what that meant. So I tried to hardcode a unicode character (the
biggest thing I could think of). That took some trial and error because a
unicode character on windows is not a wchar_t type, it is actually a char32_t
type. On windows wchar_t is only 16 bits. This suprised me.

So I finally got the character encoded and tried running my program in the emacs
shell and it came up blank. Then I tried using streams to print it, and that
degraded the unicode character down to it's decimal representation.

As I'm writing this I realize I want to try and see the binary of the unicode
character and I also want to explore running the executable in a different
terminal and see if that prints out my beer emoji.