Byte Sizes of Various Types
The output on my machine was

 char: 1
 short: 2
 int: 4
 long: 4
 long long: 8

Why are int and long the same size? That doesn't make sense to me. Shouldn't
long be bigger than int? I asked ChatGPT.

It seems that on 32 bit systems long and int are the same size. My machine, with
is a 64 bit system, maintains this trend for backwards compatibility. So 64 bit
systems are keeping long and int the same length so that they are compatible
with programs written for 32 bit systems that assume these lengths. But then
wait, why was long introduced in the first place?

It turns out it seems to be all relative. It seems that originally int was 16
bits (which makes sense). And so a long being 32 bits was in fact long(er). So
in short it seems I should use long when I want to signal to a code reader that
the value I'm trying to represent will get big. When I want to future proof my
code to rub better when larger bit systems become more ubiquitous. And when I
want my code to be portable and take advantage of systems where long is in fact
longer than int.