An interesting Footgun?
This is another example of something that I have not run into with other
programming languages.

Here, because char holder[] is assigned to an int, it is not a null terminated
string ( such as { 'h', 'i', '\0'}). printf is expecting a null terminated
string. So when it starts reading holder, it keeps reading until it hits a null
byte. This leads to undefined behavior, or in my case, the letter a being
printed twice.

This code compiles. But there is a compiler warning of a "narrowing
conversation" of 'letter' from 'int' to 'char'.