Using GDB to Manipulate Memory
My program prints out 'aa' because holder is not explicitly null terminated. So
I thought, can I get it to print out move letters if I manipulate the memory
while it's running?

And the answer is yes!

Here is how I did that. I compiled the program with the -g flag so that it would
have debugger hooks. Then I started it with `gdb chapter_code.exe`.

I set a breakpoint at the print function with `break 20`. Then I ran it with
`run`

I examined the local variables with `info locals`:
Command: info locals
letter = 97
holder = "a"

Then I looked at the address of the start of the holder variable
array. (remember this is a sneaky non null erminated array).

Command: print &holder
$1 = (char (*)[1]) 0x5ffe9b

Then I looked at the 10 addresses after the &holder array start.

x/10cb &holder
0x5ffe9b:       97 'a'  97 'a'  0 '\000'        0 '\000'        0 '\000'        32 ' '  -1 'ÿ'  95 '_'
0x5ffea3:       0 '\000'        0 '\000'

There are the two a's next to eachother followd by a null (which printf is using
as it's stopping point).

So I changed the null to a 'b'. The null is two address spaces down from the
address 0x5ffe9b. At first I wasn't sure how much I should be adding to
0x5ffe9b. I thought maybe 4 because char types are 4 bits? but it turns out I
don't fully understand memory addressing because really each of these letters is
just one address more than the next.

What does that mean?

That means that 'a' is at 0x5ffe9b and the second 'a' is at 0x5ffe9c and the
'\0' is at 0x5ffe9d. So the command to set the null terminator to a b would be

set{char}0x5ffe9d = 98

or

set{char}0x5ffe9d = 0x64 (if you want to write it in hex).