Playing with Unions
I think maybe there is potential of unions being a really good teaching tool for
myself. For example if I want to see the underlying decimal values of a variable
I can make a union and have one of the values be int and that can act as a sort
of filter over the other type (maybe a sting or whatever).