Reading and Writing to a pointer
Listing 3-2 from the book. But then I added my own little twist and
change the value assigned to *gettysburg_address to be much larger
than and int. The program compiles with a warning, and the output
loops around! This was interesting. The type checking guardrails falls
off once you start assigning values to *pointers.