Pointer decay + Pointer Arithmetic
I am still not totally sure when to use something like colleges + 2 (to get the third college in a list) and when I have to then wrap that in a dereferences *? Maybe as I'm typing it out that makes sense. Actually playing around with moving stars around and using dont notation vs arrow notation is a very good exercise.