Created A Linked List
I think I first encountered linked list in a python course in
College. The professor made them seem scary and I think that notion
has stuck with me. But I felt really good about this little
implementation. It doesn't have a lot of features, but I got the
general construction on my own.

One place I did struggle here was again passing the prefix into the
constructor of the element. I really wanted to be able to use the
shortcut of prefix{prefix} and just have that work. But, as I should
probably have already internalized, I needed to copy the string.

This was listing 3-9.