Checking out References and Unseating
Even after reading this section about "unseating" once, then again and
making a flashcard about it, and then AGAIN when I typed out the code
I still had to do a double take to remind myself what was weird about
all this.

I kept expecting he original_ref to just "not change". Instead it more
accurately 'propagates' the changes back up and into the 'original'
variable name. When I think about it, it makes sense. But the trouble
with speaking and programing is I don't take a lot of time to think
before writing things. So the trick is building that intuition.