Using Initializer Lists
I was not expecting the assignment of year_of_apert to fall down to
the year memeber of the ClockOfTheLongNow struct. But I suppose
because there is only one memeber that is public in Struct it does get
assigned by default. I wouldn't expect this to work from an intuition
standpoint.