3-4 Using Auto
I'm not totally sure how to use printf with auto. At this point it
sems like I still need to declare a formatter. I suppose with cout and
cin eventually I won't need to do that and I can really take advantage
of auto? But for now I'll just have to manuall change it from %d to %s
or whatever.