Experimenting with Static Rat-things(4-1 thru 4-3)
The book is begining to talk about static variables. This does feel
familiar from my C# days and I can imagine when a static variable
would be useful.

I think something I'm going to try and remember and watch out for is
the differences between the different scopes of static class memebers.