Exploration of Marking Functions noexcept
There is some interesting nonsense here. The book states "You can, and
should, make any function that cannot possibly throw an exception
noexcept". and follows that up a little later with "if your code throws
an exception inside a function marked no except, it's bad juju.".

What? So I made a program that did that. And sure enough if you throw
in a noexcept function, the program does not do things like
deconstruct static classes or really anything else. it just quits. (as
seen in my chapter 4 chapter_code.cpp file.

Seems risky, so what are the benefits? GPT and the book seemed to
indicate that the resulting assembly would be more efficient for
functions marked noexcept. So I generated the assembly in the
exceptions.s file. And then I was like "what is this??" and asked
GPT. And it pointed to the parts the start out _Z11with_exceptv: and
Z14with_no_exceptv:.

But I'm looking at the assembly of these functions and feeling like a
fool because the with_no_except function assembly is LONGER. So I'm like - whats up with
that??

And it turns out that the compiler is smart enough to know that I'm
literally calling a function that can throw inside my noexcept
function, and so it goes ahead and wraps in the nessicary assembly
steps to terminate. (btw, what is .cfi_personality? does my assembly
have a personality? is this a joke?)

Which now I'm like, if the compiler can see I'm about to shoot myself
in the foot why is it literally loading the gun and handing it over?
Which made me laugh. long story short, the noexcept keyword only helps
if the compiler can PROVE that you won't throw an exception. Otherwise
it adds more assembly steps.