noexcept inside Deconstructors (listing 4-13)
In c++11 destructors default to 'noexcept'. I think perhapse it is an
underappreciate fact that in most used programming languages there is
much care given to how to handle when things go wrong. And yet, the
impossibility of something going wrong feels so certain when you are
programming. It is hard to think about all the cases you have not
thought about :)