Bitwise Copy and What makes a POD a POD
My book tells me that fundamental and POD types essentially have a
bitwise copy from one memory address to another when they are called
upon in a function.

I just learned that I can determine if something is a POD vs a fancy
class by using the std::is_pod<MyStruct>::value function! which is
cool. But some other ways you can tell by looking at the code is a
class will have a userdefined constructor, virtual methods or
inheritance, private or protexted non static members, a "non-trivial"
destructor and or a base class that isn't a pod.

Which is kinda a lot of cases to remember.