Copy Assignment Operator
Another way you can make a copy to an object is to make a new object
and assign (with =) it to the first object. Like in line 49 where I
say p2 = p1. But the book warns against this because you still have to
clean up p2's resources and if the class owns dynamically allocated
memory, such as a 'new char[]', that array is nto freed.

Additionally, if you copy like this, both classes will own the
dynamically allocated memory (which can cause double frees and
dangling pointers).

The lesson is that you must implement a copy assignment operator that
performs a clean hand-off.