Finished Chapter 4 Exercises
I didn't do number 7 because I'm going to just include that in my
flash cards. But I found exercises 1 through 6 to be realyl good! I
need to play around with this some more, maybe come back tomorrow and
try it again from scratch. I am not used to having to implement things
that are not explicitly asked for by the compiler.

Also I'm still not totally comfortable with the move constructor and
move assignment syntax. But I'm also almost a quarer way through the
book, which is exciting!