Introducing My New Build System: SCons
I found SCons because the Godot project uses it to generate build
files as well as IDE files for building Godot from scource. This kind
of cross platform build reproducibility feels really important!

So I started reading about Cmake and a from mentioned Meson and also I
see you Autotools. I ran cmake and it generated truly a billion files,
which felt overwhelming. So I deleted all those files and lookd into
scons (my experience of getting Godot set up on my machine was
actually pretty painless considering how massive that program is).

I was really drawn to the fact that SCons uses python as it's
scripting language as opposed to some bespoke situation of yaml or
JSON. Also it didn't generate a billion files. And I it was pretty
easy to tell it to put all the files it did generate into a build
directory.

I think my current pattern here is going to be that for each chapter I
include a new SConstruct.py file.

Additionally, I was able to use Scons in a different project (my maze)
and upload it to github, and from there use github actions to
basically cross compile the whole thing which was very cool! and kind
of a mini objective for myself.