Inheritance And Interfaces
I'm excited to use this in the wild. This concept feels pretty
familiar coming from C#. I also messed around in the chapter05 code
section and got cmake working as another optional build framework. At
first my approach was one or the other, but I think learning a little
of both could be interesting to see where the tradeoffs are. For
example for simple projects cmake kinda requires a bunch of
definitions. But SCons has a super minimal build as well as compile
command that I really enjoy.