Listing 6-11
I took a bit of a break and it feels good to be back! Templates feel
like an understandable concept. Very similar to generics in C# so
feeling strong there. It is interesting that one problem I ran into
was that passing a result into print you don't actually know what type
it is.

I guess to solve this I could wrap printf in my own template function
that would have a switch statement that would use the correct
%(string templating function).