Listing 6-14: A simple Unique Pointer Implementation
In my flashcards I have started memorizing the 5 basic
constructors/move assignments for the SimpleString class and that has
been really helpful here. I didn't need to look up what all these
methods were doing or when they would be called. I was able to
remember which made this easier to understand.

I think I will also try and memorize/review this as well because I
think it is a really book basic implementation that touches on a lot
of core C++ ideas.