Listing 6-21, using concepts
Concepts pairs with requirements and constraints. So one technical
thing here is that float += float does not return float, but instead
float&.

This is a cool subtlety and one that I honestly don't totally
understand right now. Basically it is because the result of lhs += rhs
is a reference to the left-hand operand, not a new float!

A += b modifies a in place, and returns a reference to a. This allows
for chaining.

a += b += c

If this returned a value, you could not chain like this.

ChatGPT wants me to image that because this statement returns a
reference to a (in other words &a), I can continue to "pour things into
the box". Which I think for right now is just something I'm going to
have to memorize rather than perfectly understand.