Exercise 6-1, Mode Toad
I know this is the perfect use for a lill hash table, but I wanted to
solve it with the tools the book has already given me. And I'm pretty
proud of how I did.

Specifically, at first in the mode function I dynamically allocated a
Frequencies struct with `Frequencies f = new Frequencies(length)` and
then I realized afterwards that I don't need the Frequencies struct to
be on the heap because I don't need it to exist beyond scope. I do
think that I should create a deconstructor for the Frequencies object
so that the lower case frequencies array is properly deleted as the
program moves on.

This took me about 45 minutes to write. So we are getting there.